executor double hiba

+++ src/main/java/hu/meiit/haladojava/calculator/Executor.java
@@ -2,7 +2,7 @@ package hu.meiit.haladojava.calculator;
 
 public class Executor {
     public double add(int operand1, int operand2){
-        return operand1+operand2;
+        return (double)operand1+(double)operand2;
     }
     public double sub(int operand1, int operand2){
         return operand1-operand2;
